Cleans up init by separating windowing and frame dumper
