+++ crates/bitang/src/file/chart_file.rs
@@ -5,8 +5,8 @@ use crate::loader::resource_repository::ResourceRepository;
 use crate::loader::ResourcePath;
 use crate::render::buffer_generator::BufferGeneratorType;
 use crate::render::image::ImageSizeRule;
-use crate::render::vulkan_window::VulkanContext;
 use crate::render::SCREEN_RENDER_TARGET_ID;
+use crate::tool::VulkanContext;
 use crate::{file, render};
 use ahash::AHashMap;
 use anyhow::{anyhow, Context, Result};
