+++ crates/bitang/src/file/project_file.rs
@@ -1,6 +1,6 @@
 use crate::loader::resource_repository::ResourceRepository;
 use crate::render;
-use crate::render::vulkan_window::VulkanContext;
+use crate::tool::VulkanContext;
 use anyhow::Result;
 use futures::future::join_all;
 use serde::Deserialize;
