+++ crates/bitang/src/loader/resource_cache.rs
@@ -1,11 +1,11 @@
 use crate::loader::async_cache::{AsyncCache, LoadFuture};
 use crate::loader::file_cache::{ContentHash, FileCache, FileCacheEntry};
 use crate::loader::ResourcePath;
-use crate::render::vulkan_window::VulkanContext;
+use crate::tool::VulkanContext;
 use anyhow::Result;
 use std::sync::Arc;
 use tokio::task::spawn_blocking;
-use tracing::{trace};
+use tracing::trace;
 
 type LoaderFunc<T> =
     fn(context: &Arc<VulkanContext>, blob: &[u8], resource_name: &str) -> Result<Arc<T>>;
