+++ crates/bitang/src/render/chart.rs
@@ -4,7 +4,7 @@ use crate::render::buffer_generator::BufferGenerator;
 use crate::render::camera::Camera;
 use crate::render::draw::Draw;
 use crate::render::image::Image;
-use crate::render::vulkan_window::RenderContext;
+use crate::tool::RenderContext;
 use anyhow::Result;
 use std::sync::Arc;
