+++ crates/bitang/src/render/draw.rs
@@ -2,7 +2,7 @@ use crate::control::controls::{Control, Globals};
 use crate::render::camera::Camera;
 use crate::render::pass::Pass;
 use crate::render::render_object::RenderObject;
-use crate::render::vulkan_window::RenderContext;
+use crate::tool::RenderContext;
 use anyhow::{ensure, Result};
 use glam::{Mat4, Vec2, Vec3};
 use std::sync::Arc;
