+++ crates/bitang/src/render/mesh.rs
@@ -1,5 +1,5 @@
-use crate::render::vulkan_window::VulkanContext;
 use crate::render::Vertex3;
+use crate::tool::VulkanContext;
 use anyhow::Result;
 use std::sync::Arc;
 use vulkano::buffer::{Buffer, BufferCreateInfo, BufferUsage, Subbuffer};
@@ -15,7 +15,7 @@ pub struct Mesh {
 impl Mesh {
     pub fn try_new(context: &Arc<VulkanContext>, vertices: Vec<Vertex3>) -> Result<Mesh> {
         let vertex_buffer = Buffer::from_iter(
-            context.vulkano_context.memory_allocator(),
+            &context.memory_allocator,
             BufferCreateInfo {
                 usage: BufferUsage::VERTEX_BUFFER,
                 ..Default::default()
